added chardin.js overlays to show info
Once the page is centered in the document (which has been frustrating to
implement), the overlay info can be placed on the left or right
depending on the location of the bounding box.

Also, the image itself probably needs to be returned in base64 form so
that when the overlay is rendered, it doesn't darken the image itself.
It darkens it now because the image is just part of the PDF render and
not a separate element