removed page skipping due to PDFUnicodeNotDefined
I was getting this error on a couple pages for unknown reasons, but it
was affecting a page with a watermarked image, leading to that page
being skipped, and that image not being highlighted in the viewer.
Looking into the error I was getting, it only happened when pdfminer was
supposed to translate unicode characters 48 and 88 on pages 6 and 7
specifically. So I hacked pdfminer/pdffont.py to just return u'0' and
u'X', respectively and this got rid of all the problems